Refactor: 로컬 상태 일부를 전역 상태로 이전하고 컴포넌트 역할 재정의
#31
