+++ src/store/useSettingStore.jsx
@@ -2,8 +2,10 @@ import { create } from "zustand";
 
 const useSettingStore = create((set) => ({
   groundType: "wood_dark",
+  rotationSensitivity: 1,
 
-  setGroundType: (type) => set(() => ({ groundType: type })),
+  setGroundType: (groundType) => set(() => ({ groundType })),
+  setRotationSensitivity: (sensitivity) => set(() => ({ rotationSensitivity: sensitivity })),
 }));
 
 export default useSettingStore;
